Create Simple MDP for RPS
This program draws inspiration from Kaggle's competition for Rock, Paper Scissors. The idea is that humans cannot exhibit random behavior, and always follow a specific pattern, and an agent should be able to beat them when playing a large number of games. 
This code uses the Markov Property (assuming that it holds) and uses historical data to predict the user's next move. 
Code can be adjusted so a human can play instead of the "user".